requirements.txt
# المتطلبات الأساسية
telethon>=1.34.0
python-dotenv>=1.0.0

# قاعدة البيانات (مدمجة مع Python)
# sqlite3 - مدمج

# دوال مساعدة
aiofiles>=23.2.0
asyncio>=3.4.3

# التشفير (اختياري)
cryptography>=41.0.0

# معالجة الصور (اختياري)
Pillow>=10.0.0